+++ Docs/Changes.md
@@ -1,0 +1,5 @@
+# 0.2.0
+wip
+
+# 0.1.0
+Initial release
